+++ src/main/java/de/chrgroth/smartcron/model/SmartcronMetadata.java
@@ -12,17 +12,22 @@ import java.util.List;
 public class SmartcronMetadata {
 
     private final String name;
-    private final List<SmartcronExecution> executions;
+    private final List<SmartcronExecution> executions = new ArrayList<>();
+    private final long avgDuration;
+    private final double errorPercentage;
     private final LocalDateTime scheduled;
 
     public SmartcronMetadata(String name, List<SmartcronExecution> executions, LocalDateTime scheduled) {
         this.name = name;
-        this.executions = new ArrayList<>();
-        if (executions != null) {
+        if (executions != null && !executions.isEmpty()) {
             this.executions.addAll(executions);
+            avgDuration = Math.round((double) executions.stream().mapToLong(se -> se.getDuration()).sum() / (double) executions.size());
+            errorPercentage = Math.round((double) executions.stream().filter(se -> se.getError() != null).count() / (double) executions.size() * 100) / 100;
+        } else {
+            avgDuration = 0;
+            errorPercentage = 0;
         }
         this.scheduled = scheduled;
-
     }
 
     public String getName() {
@@ -30,7 +35,15 @@ public class SmartcronMetadata {
     }
 
     public List<SmartcronExecution> getExecutions() {
-        return executions;
+        return new ArrayList<>(executions);
+    }
+
+    public long getAvgDuration() {
+        return avgDuration;
+    }
+
+    public double getErrorPercentage() {
+        return errorPercentage;
     }
 
     public LocalDateTime getScheduled() {
@@ -39,6 +52,7 @@ public class SmartcronMetadata {
 
     @Override
     public String toString() {
-        return "SmartcronMetadata [name=" + name + ", executions=" + executions + ", scheduled=" + scheduled + "]";
+        return "SmartcronMetadata [name=" + name + ", executions=#" + executions.size() + ", avgDuration=" + avgDuration + ", errorPercentage=" + errorPercentage + ", scheduled="
+                + scheduled + "]";
     }
 }